backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.14.1
pandas==2.3.2
scikit-learn==1.7.1
joblib==1.5.1
python-dotenv==1.1.1
